[어드민 프로젝트] 상품 API CREATE, READ, UPDATE, DELETE 구현

+++ src/main/java/com/example/study/model/entity/Item.java
@@ -1,5 +1,6 @@
 package com.example.study.model.entity;
 
+import ch.qos.logback.classic.turbo.TurboFilter;
 import lombok.*;
 import lombok.experimental.Accessors;
 import org.springframework.data.annotation.CreatedBy;
@@ -9,6 +10,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -34,7 +36,7 @@ public class Item {
 
     private String content;
 
-    private Integer price;
+    private BigDecimal price;
 
     private String brandName;
